Move FileUploadComponent to test folder as base upload component
