+++ src/Features/SupportFileUploads/UnitTest.php
@@ -16,6 +16,7 @@ use Illuminate\Http\UploadedFile;
 use Illuminate\Http\Testing\FileFactory;
 use Illuminate\Support\Arr;
 use Livewire\Attributes\Validate;
+use Tests\FileUploadComponent;
 use Tests\TestComponent;
 
 class UnitTest extends \Tests\TestCase
@@ -1289,88 +1290,6 @@ class DummyMiddleware
 class NonFileUploadComponent extends TestComponent
 {
     public $photo;
-}
-
-class FileUploadComponent extends TestComponent
-{
-    use WithFileUploads;
-
-    public $file;
-    public $photo;
-    public $photos;
-    public $photosArray = [];
-    public $storedFilename;
-
-    public function updatedPhoto()
-    {
-        $this->validate(['photo' => 'image|max:300']);
-    }
-
-    public function updatedPhotos()
-    {
-        $this->validate(['photos.*' => 'image|max:300']);
-    }
-
-    public function upload($name)
-    {
-        $this->photo->storeAs('/', $name, $disk = 'avatars');
-    }
-
-    public function uploadMultiple($baseName)
-    {
-        $number = 1;
-
-        foreach ($this->photos as $photo) {
-            $photo->storeAs('/', $baseName.$number++.'.png', $disk = 'avatars');
-        }
-    }
-
-    public function uploadPhotosArray($baseName)
-    {
-        $number = 1;
-
-        foreach ($this->photosArray as $photo) {
-            $photo->storeAs('/', $baseName.$number++.'.png', $disk = 'avatars');
-        }
-    }
-
-    public function uploadAndSetStoredFilename()
-    {
-        $this->storedFilename = $this->photo->store('/', $disk = 'avatars');
-    }
-
-    public function validateUpload()
-    {
-        $this->validate(['photo' => 'file|max:100']);
-    }
-
-    public function validateMultipleUploads()
-    {
-        $this->validate(['photos.*' => 'file|max:100']);
-    }
-
-    public function validateUploadWithDimensions()
-    {
-        $this->validate([
-            'photo' => Rule::dimensions()->maxWidth(100)->maxHeight(100),
-        ]);
-    }
-
-    public function removePhoto($key) {
-        unset($this->photos[$key]);
-    }
-
-    public function uploadError($name)
-    {
-        $this->_uploadErrored($name, null, false);
-    }
-
-    public function uploadErrorWithMalformedJson($name)
-    {
-        // Simulate malformed JSON without 'errors' key
-        $malformedJson = '{"message":"Something went wrong"}';
-        $this->_uploadErrored($name, $malformedJson, false);
-    }
 }
 
 class FileUploadInArrayComponent extends FileUploadComponent
